make edit bestuursorgaan popup
