+++ app/router.js
@@ -29,7 +29,6 @@ Router.map(function () {
   this.route('organen', function () {
     this.route('orgaan', { path: '/:id' }, function () {
       this.route('mandatarissen');
-      this.route('edit');
       this.route('mandataris', function () {
         this.route('new');
       });
