--- app/templates/organen/orgaan/edit.hbs
@@ -1,25 +1,0 @@
-{{page-title "Bewerk Bestuursorgaan" " " @model.bestuursorgaan.naam}}
-
-<AuToolbar @border="bottom" @size="large" as |Group|>
-  <Group>
-    <AuHeading @skin="2">Bewerk bestuursorgaan</AuHeading>
-  </Group>
-</AuToolbar>
-
-<div class="au-c-main-container">
-  <div class="au-o-box">
-    <Form::Instance
-      @instanceId={{this.model.instanceId}}
-      @form={{this.model.form}}
-      @onSave={{this.closeForm}}
-    />
-  </div>
-</div>
-
-<AuToolbar @border="top" @size="large" as |Group|>
-  <Group>
-    <Form::CancelWithConfirm
-      @cancel={{this.closeForm}}
-    >Annuleer</Form::CancelWithConfirm>
-  </Group>
-</AuToolbar>
